+++ opencv/src/main/java/org/bytedeco/javacpp/opencv_highgui.java
@@ -1,4 +1,4 @@
-// Targeted by JavaCPP version 0.10
+// Targeted by JavaCPP version 0.11-SNAPSHOT
 
 package org.bytedeco.javacpp;
 
@@ -304,6 +304,8 @@ public static native int cvGetTrackbarPos( @Cast("const char*") BytePointer trac
 public static native int cvGetTrackbarPos( String trackbar_name, String window_name );
 public static native void cvSetTrackbarPos( @Cast("const char*") BytePointer trackbar_name, @Cast("const char*") BytePointer window_name, int pos );
 public static native void cvSetTrackbarPos( String trackbar_name, String window_name, int pos );
+public static native void cvSetTrackbarMax(@Cast("const char*") BytePointer trackbar_name, @Cast("const char*") BytePointer window_name, int maxval);
+public static native void cvSetTrackbarMax(String trackbar_name, String window_name, int maxval);
 
 /** enum  */
 public static final int
@@ -563,7 +565,7 @@ public static final int
     CV_CAP_PROP_GAIN          = 14,
     CV_CAP_PROP_EXPOSURE      = 15,
     CV_CAP_PROP_CONVERT_RGB   = 16,
-    CV_CAP_PROP_WHITE_BALANCE_BLUE_U = 17,
+    CV_CAP_PROP_WHITE_BALANCE_U = 17,
     CV_CAP_PROP_RECTIFICATION = 18,
     CV_CAP_PROP_MONOCROME     = 19,
     CV_CAP_PROP_SHARPNESS     = 20,
@@ -574,7 +576,7 @@ public static final int
     CV_CAP_PROP_TEMPERATURE   = 23,
     CV_CAP_PROP_TRIGGER       = 24,
     CV_CAP_PROP_TRIGGER_DELAY = 25,
-    CV_CAP_PROP_WHITE_BALANCE_RED_V = 26,
+    CV_CAP_PROP_WHITE_BALANCE_V = 26,
     CV_CAP_PROP_ZOOM          = 27,
     CV_CAP_PROP_FOCUS         = 28,
     CV_CAP_PROP_GUID          = 29,
@@ -586,6 +588,7 @@ public static final int
     CV_CAP_PROP_ROLL          = 35,
     CV_CAP_PROP_IRIS          = 36,
     CV_CAP_PROP_SETTINGS      = 37,
+    CV_CAP_PROP_BUFFERSIZE    = 38,
 
     CV_CAP_PROP_AUTOGRAB      = 1024, // property for highgui class CvCapture_Android only
     CV_CAP_PROP_SUPPORTED_PREVIEW_SIZES_STRING= 1025, // readonly, tricky property, returns cpnst char* indeed