+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_calib3d.java
@@ -32,7 +32,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  */
 @Properties(inherit={opencv_highgui.class, opencv_features2d.class}, value={
     @Platform(include="<opencv2/calib3d/calib3d.hpp>", link="opencv_calib3d@.2.4"),
-    @Platform(value="windows", link="opencv_calib3d2410") },
+    @Platform(value="windows", link="opencv_calib3d2411") },
         target="org.bytedeco.javacpp.opencv_calib3d", helper="org.bytedeco.javacpp.helper.opencv_calib3d")
 public class opencv_calib3d implements InfoMapper {
     public void map(InfoMap infoMap) {