+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_features2d.java
@@ -32,7 +32,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  */
 @Properties(inherit={opencv_highgui.class, opencv_flann.class}, value={
     @Platform(include="<opencv2/features2d/features2d.hpp>", link="opencv_features2d@.2.4"),
-    @Platform(value="windows", link="opencv_features2d2410") },
+    @Platform(value="windows", link="opencv_features2d2411") },
         target="org.bytedeco.javacpp.opencv_features2d")
 public class opencv_features2d implements InfoMapper {
     public void map(InfoMap infoMap) {