+++ opencv/src/main/java/org/bytedeco/javacpp/presets/opencv_imgproc.java
@@ -32,7 +32,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  */
 @Properties(inherit=opencv_core.class, value={
     @Platform(include={"<opencv2/imgproc/types_c.h>", "<opencv2/imgproc/imgproc_c.h>", "<opencv2/imgproc/imgproc.hpp>"}, link="opencv_imgproc@.2.4"),
-    @Platform(value="windows", link="opencv_imgproc2410") },
+    @Platform(value="windows", link="opencv_imgproc2411") },
         target="org.bytedeco.javacpp.opencv_imgproc", helper="org.bytedeco.javacpp.helper.opencv_imgproc")
 public class opencv_imgproc implements InfoMapper {
     public void map(InfoMap infoMap) {