+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 
+ * Let `createBuffer()` return `UByteIndexer` and `UShortIndexer` when appropriate for unsigned data types
  * Remove the need to set manually the `platform.dependency` system property for downstream modules without a `cppbuild.sh` file
  * Fix failing `cppbuild.sh` for FFmpeg on Ubuntu ([issue #32](https://github.com/bytedeco/javacpp-presets/issues/32))
  * Disable iconv, XCB, and SDL for more portables builds of FFmpeg
  * Avoid versioning of Leptonica and Tesseract libraries, preventing them from working on Android ([issue #38](https://github.com/bytedeco/javacpp-presets/issues/38))
  * Add x265 in the `cppbuild.sh` script for FFmpeg, thanks to Mark Bolstad ([issue bytedeco/javacv#41](https://github.com/bytedeco/javacv/issues/41))
- * Upgrade presets for FFmpeg 2.5.4, OpenSSL 1.0.2
+ * Upgrade presets for OpenCV 2.4.11, FFmpeg 2.5.4, OpenSSL 1.0.2
  * Switch from `IntPointer` to `BoolPointer` for the `BOOL*` pointer type of Leptonica ([issue #36](https://github.com/bytedeco/javacpp-presets/issues/36))
  * Add `preload` for `gif`, `jpeg`, `png`, `tiff`, and `webp` libraries in presets for Leptonica ([issue #36](https://github.com/bytedeco/javacpp-presets/issues/36))
  * Include missing `ltrresultiterator.h` header file in the presets for Tesseract ([issue #36](https://github.com/bytedeco/javacpp-presets/issues/36))